fix(module): fix name and implement validation
