--- src/app/Components/List/list.component.html
@@ -1,5 +1,0 @@
-
-  <div class="row">
-    <app-dragon-card (reloaderDragons) ="reloadDragons($event)" [dragon]="dragon" class="col-sm-12 col-md-4 col-lg-4 mx-auto" *ngFor="let dragon of dragons;">
-    </app-dragon-card>
-  </div>
