+++ src/app/Components/list/list.component.scss
@@ -1,5 +1,9 @@
 @import "../../../variables.scss";
 
+h1{
+  color: $second-color;
+}
+
 .container-list{
   background-color: black;
 }
