+++ src/app/Components/list/list.module.ts
@@ -1,11 +1,16 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { ListComponent } from './list.component';
+import { AppRoutingModule } from 'src/app/app-routing.module';
 
 
 
 @NgModule({
-  declarations: [],
+  declarations: [
+    ListComponent
+  ],
   imports: [
+    AppRoutingModule,
     CommonModule
   ],
   providers: [
